8: moved App.jsx to components folder and added routing
