+++ package.json
@@ -43,6 +43,7 @@
     "react": "^16.3.1",
     "react-dev-utils": "^5.0.1",
     "react-dom": "^16.3.1",
+    "react-router-dom": "^4.2.2",
     "resolve": "1.6.0",
     "sass-loader": "^6.0.7",
     "style-loader": "0.19.0",
